requirements.txt
# Core Framework
streamlit==1.31.0
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5

# WebSocket
websocket-client==1.7.0
python-socketio==5.11.0

# Data Processing
pandas==2.1.4
numpy==1.26.3
scipy==1.11.4

# Analytics
statsmodels==0.14.1
scikit-learn==1.4.0

# Visualization
plotly==5.18.0
matplotlib==3.8.2

# Storage
redis==5.0.1
sqlalchemy==2.0.25

# Utilities
python-dotenv==1.0.0
requests==2.31.0
